✨ create AggregateRecords/ScanRecords extension methods
