+++ SurrealDb.Net.LiveQuery.Tests/ReactiveLiveQueryTests.Subscribe.cs
@@ -35,7 +35,7 @@ public class ReactiveLiveQueryTests : BaseLiveQueryTests
             await using var liveQuery = client.ListenLive<TestRecord>(liveQueryUuid);
 
             var cts = new CancellationTokenSource();
-            cts.CancelAfter(TimeSpan.FromSeconds(2));
+            cts.CancelAfter(Timeout);
 
             var testScheduler = new TestScheduler();
 
@@ -98,7 +98,7 @@ public class ReactiveLiveQueryTests : BaseLiveQueryTests
             var liveQuery = client.ListenLive<TestRecord>(liveQueryUuid);
 
             var cts = new CancellationTokenSource();
-            cts.CancelAfter(TimeSpan.FromSeconds(2));
+            cts.CancelAfter(Timeout);
 
             var testScheduler = new TestScheduler();
 
@@ -167,7 +167,7 @@ public class ReactiveLiveQueryTests : BaseLiveQueryTests
             await using var liveQuery = client.ListenLive<TestRecord>(liveQueryUuid);
 
             var cts = new CancellationTokenSource();
-            cts.CancelAfter(TimeSpan.FromSeconds(2));
+            cts.CancelAfter(Timeout);
 
             var record = await client.Create("test", new TestRecord { Value = 1 });
             await WaitLiveQueryNotificationAsync();
@@ -194,118 +194,5 @@ public class ReactiveLiveQueryTests : BaseLiveQueryTests
 
         var firstResult = allResults[0];
         firstResult.Should().BeOfType<SurrealDbLiveQueryDeleteResponse>();
-    }
-
-    [Fact]
-    public async Task ShouldConsumeAggregatedData()
-    {
-        const string url = "ws://localhost:8000/rpc";
-
-        List<TestRecord>? records = null;
-
-        Func<Task> func = async () =>
-        {
-            await using var surrealDbClientGenerator = new SurrealDbClientGenerator();
-            var dbInfo = surrealDbClientGenerator.GenerateDatabaseInfo();
-
-            using var client = surrealDbClientGenerator.Create(url);
-            await client.SignIn(new RootAuth { Username = "root", Password = "root" });
-            await client.Use(dbInfo.Namespace, dbInfo.Database);
-
-            var response = await client.Query("LIVE SELECT * FROM test;");
-
-            if (response.FirstResult is not SurrealDbOkResult okResult)
-                throw new Exception("Expected a SurrealDbOkResult");
-
-            var liveQueryUuid = okResult.GetValue<Guid>();
-
-            var liveQuery = client.ListenLive<TestRecord>(liveQueryUuid);
-
-            var cts = new CancellationTokenSource();
-            cts.CancelAfter(TimeSpan.FromSeconds(2));
-
-            var testScheduler = new TestScheduler();
-
-            using var _ = liveQuery
-                .GetResults()
-                .ToObservable()
-                .Aggregate(
-                    new Dictionary<string, TestRecord>(), // Start from an empty list or from the current list of records
-                    (acc, response) =>
-                    {
-                        if (
-                            response is SurrealDbLiveQueryCreateResponse<TestRecord> createResponse
-                            && createResponse.Result.Id is not null
-                        )
-                        {
-                            acc[createResponse.Result.Id.ToString()] = createResponse.Result;
-                        }
-                        if (
-                            response is SurrealDbLiveQueryUpdateResponse<TestRecord> updateResponse
-                            && updateResponse.Result.Id is not null
-                        )
-                        {
-                            acc[updateResponse.Result.Id.ToString()] = updateResponse.Result;
-                        }
-                        if (
-                            response is SurrealDbLiveQueryDeleteResponse deleteResponse
-                            && deleteResponse.Result is not null
-                        )
-                        {
-                            acc.Remove(deleteResponse.Result.ToString());
-                        }
-
-                        return acc;
-                    }
-                )
-                .Select(x => x.Values.ToList())
-                .SubscribeOn(testScheduler)
-                .Subscribe(results =>
-                {
-                    records = results;
-                });
-
-            await WaitLiveQueryCreationAsync();
-
-            testScheduler.Start();
-
-            var record1 = await client.Create("test", new TestRecord { Value = 1 });
-            await WaitLiveQueryNotificationAsync();
-
-            await client.Upsert(new TestRecord { Id = record1.Id, Value = 2 });
-            await WaitLiveQueryNotificationAsync();
-
-            var record2 = await client.Create("test", new TestRecord { Value = 44 });
-            await WaitLiveQueryNotificationAsync();
-
-            var record3 = await client.Create("test", new TestRecord { Value = 66 });
-            await WaitLiveQueryNotificationAsync();
-
-            var record4 = await client.Create("test", new TestRecord { Value = 8 });
-            await WaitLiveQueryNotificationAsync();
-
-            await client.Upsert(new TestRecord { Id = record1.Id, Value = 2 });
-            await WaitLiveQueryNotificationAsync();
-
-            await client.Upsert(new TestRecord { Id = record2.Id, Value = 45 });
-            await WaitLiveQueryNotificationAsync();
-
-            await client.Delete(record1.Id!);
-            await WaitLiveQueryNotificationAsync();
-
-            await liveQuery.DisposeAsync();
-            await WaitLiveQueryNotificationAsync();
-        };
-
-        await func.Should().NotThrowAsync();
-
-        var expected = new List<TestRecord>
-        {
-            new() { Id = records![0].Id, Value = 45 },
-            new() { Id = records![1].Id, Value = 66 },
-            new() { Id = records![2].Id, Value = 8 },
-        };
-
-        records.Should().BeEquivalentTo(expected);
     }
 }
