+++ SurrealDb.Net.LiveQuery.Tests/SurrealDb.Net.LiveQuery.Tests.csproj
@@ -28,6 +28,7 @@
   <ItemGroup>
     <ProjectReference Include="..\SurrealDb.Net.Tests.Fixtures\SurrealDb.Net.Tests.Fixtures.csproj" />
     <ProjectReference Include="..\SurrealDb.Net\SurrealDb.Net.csproj" />
+    <ProjectReference Include="..\SurrealDb.Reactive\SurrealDb.Reactive.csproj" />
   </ItemGroup>
 
 </Project>
