+++ SurrealDb.Reactive/SurrealDb.Reactive.csproj
@@ -1,0 +1,31 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.1;net6.0;net7.0</TargetFrameworks>
+    <LangVersion>latest</LangVersion>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+
+    <PackageId>SurrealDb.Reactive</PackageId>
+    <Version>0.1.4</Version>
+    <Authors>surrealdb</Authors>
+    <Description>Reactive Extensions for the SurrealDB library for .NET</Description>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <PackageProjectUrl>https://github.com/surrealdb/surrealdb.net</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/surrealdb/surrealdb.net</RepositoryUrl>
+    <Company>SurrealDB</Company>
+    <PackageIcon>icon.png</PackageIcon>
+    <PackageTags>SurrealDB Surreal Database</PackageTags>
+    <Copyright>Copyright © SurrealDB Ltd</Copyright>
+
+    <EnablePackageValidation>true</EnablePackageValidation>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <NoWarn>$(NoWarn);CS1591</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\SurrealDb.Net\SurrealDb.Net.csproj" />
+  </ItemGroup>
+
+</Project>
